Renamed default handlebars function and added base parser class
