+++ spec/fixtures/example.handlebars
@@ -1,10 +1,10 @@
 <div class="content-main">
   <div class="content-header">
-    <label>{{_ "Locale"}}</label>
-    <h1>{{_ "Profile"}}</h1>
+    <label>{{__ "Locale"}}</label>
+    <h1>{{__ "Profile"}}</h1>
     <div class="btn-toolbar pull-right">
       <div class="btn-group">
-        <a href="#" {{action "updateUser" target="controller"}} class="btn btn-success">{{_ "Update" "Updates" 2}}</a>
+        <a href="#" {{action "updateUser" target="controller"}} class="btn btn-success">{{__ "Update" "Updates" 2}}</a>
       </div>
     </div>
   </div>
